+++ changes.md
@@ -1,5 +1,11 @@
 # Changes
 
+## Devel
+
+- Qt mainloop scheduler thanks to @jdreaver.
+- Use `threading.RLock` instead of `threading.Lock` since `BehaviorSubject`
+  may share lock with "child" operator at subscribe time. Fixes #50
+
 ## 1.2.1
 
 - Fix to preserve the original error message for exceptions #44, thanks